Extract chat continuation helpers
