+++ tests/test-chat-actions.js
@@ -158,6 +158,7 @@ console.log('Section 17: Source inspection');
 const chatSrc = read('js/chat.js');
 const chatIconsSrc = read('js/chat-icons.js');
 const chatSummariesSrc = read('js/chat-summaries.js');
+const chatContinuationSrc = read('js/chat-continuation.js');
 const labCtxSrc = read('js/lab-context.js');
 assert('lab-context.js has getContextSummary', labCtxSrc.includes('function getContextSummary'), 'found');
 assert('chat.js has buildActionBar', chatSrc.includes('function buildActionBar'), 'found');
@@ -165,22 +166,24 @@ assert('chat.js has regenerateLastMessage', chatSrc.includes('function regenerat
 assert('chat.js does NOT have readAloud', !chatSrc.includes('function readAloud'), 'removed');
 assert('chat.js has copyMessage', chatSrc.includes('function copyMessage'), 'found');
 assert('sendChatMessage snapshots context', chatSrc.includes('contextSnapshot'), 'found');
-assert('chat raises response token headroom', chatSrc.includes('CHAT_RESPONSE_MAX_TOKENS = 16384'), 'found');
-assert('chat auto-continues token-limit stops', chatSrc.includes('CHAT_AUTO_CONTINUE_LIMIT') && chatSrc.includes('_callChatAPIWithContinuation'), 'found');
-assert('chat auto-continues likely mid-sentence stops', chatSrc.includes('_isLikelyIncompleteResponse') && chatSrc.includes('_shouldAutoContinueResponse'), 'found');
+assert('chat raises response token headroom', chatContinuationSrc.includes('CHAT_RESPONSE_MAX_TOKENS = 16384'), 'found');
+assert('chat auto-continues token-limit stops', chatContinuationSrc.includes('CHAT_AUTO_CONTINUE_LIMIT') && chatContinuationSrc.includes('callChatAPIWithContinuation'), 'found');
+assert('chat auto-continues likely mid-sentence stops', chatContinuationSrc.includes('isLikelyIncompleteResponse') && chatContinuationSrc.includes('shouldAutoContinueResponse'), 'found');
 assert('chat incomplete heuristic does not continue solely because final line is long',
   !chatSrc.includes('return lastLine.length > 60'), 'length-only fallback removed');
 assert('chat incomplete heuristic does not continue on terminal high/low adjectives',
   !chatSrc.includes('low|high') && !chatSrc.includes('high|low'), 'medical adjectives removed from trailing-word fallback');
-assert('chat renders output-limit note', chatSrc.includes('output limit reached'), 'found');
+assert('chat renders output-limit note', chatContinuationSrc.includes('output limit reached'), 'found');
 assert('chat persists truncated assistant state', chatSrc.includes('assistantMsg.truncated = true'), 'found');
-assert('renderChatMessages restores truncated note', chatSrc.includes('msg.truncated') && chatSrc.includes('_responseLimitNote()'), 'found');
+assert('renderChatMessages restores truncated note', chatSrc.includes('msg.truncated') && chatSrc.includes('responseLimitNote()'), 'found');
 assert('regenerateLastMessage checks _chatAbortController', chatSrc.includes('_chatAbortController') && chatSrc.includes('regenerateLastMessage'), 'found');
 assert('chat.js imports chat icon helpers', chatSrc.includes("from './chat-icons.js'"), 'found');
 assert('chat-icons.js exports button content helper', chatIconsSrc.includes('export function setIconButtonContent'), 'found');
 assert('chat.js imports chat summary helpers', chatSrc.includes("from './chat-summaries.js'"), 'found');
 assert('chat-summaries.js exports summarizeThread', chatSummariesSrc.includes('export async function summarizeThread'), 'found');
 assert('chat-summaries.js sends one transcript message', chatSummariesSrc.includes('buildSummaryTranscript(state.chatHistory)') && chatSummariesSrc.includes("role: 'user'"), 'found');
+assert('chat.js imports continuation helpers', chatSrc.includes("from './chat-continuation.js'"), 'found');
+assert('chat-continuation.js exports continuation helper', chatContinuationSrc.includes('export async function callChatAPIWithContinuation'), 'found');
 assert('renderChatMessages calls buildActionBar', chatSrc.includes('buildActionBar(i)'), 'found');
 assert('API messages tag other personas', chatSrc.includes('Response from') && chatSrc.includes('personalityName'), 'tags messages from different personas');
